Creating the first version of the source code
This file will currently calculate the velocity in the pipe when the flowrate and diameter are input by the user. 
It will also calculate the Reynolds number and decide if the flow is laminar or turbulent. For the lab it is designed for, the flow should always be turbulent is why it is hard coded the way it is.